refactor: dev.ts
